# public/api-docs.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ogami ERP API",
...
    "version": "1.0.0",
    "contact": {
      "name": "Ogami ERP Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1",
      "description": "Local development server"
    },
    {
      "url": "https://api.ogami-erp.local/api/v1",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/leave/requests": {
      "get": {
        "tags": ["Leave"],
        "summary": "List leave requests",
        "security": [{"sanctum": []}],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {"type": "string", "enum": ["pending", "approved", "rejected", "cancelled"]},
            "description": "Filter by status"
          },
          {
            "name": "employee_id",
            "in": "query",
            "schema": {"type": "integer"},
            "description": "Filter by employee"
          }
        ],
        "responses": {
          "200": {
            "description": "List of leave requests"
          }
        }
      },
      "post": {
        "tags": ["Leave"],
        "summary": "Submit leave request",
        "security": [{"sanctum": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeaveRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Leave request submitted"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LeaveRequest": {
        "type": "object",
        "properties": {
          "leave_type_id": {"type": "integer"},
          "start_date": {"type": "string", "format": "date"},
          "end_date": {"type": "string", "format": "date"},
          "reason": {"type": "string"},
          "half_day": {"type": "boolean", "default": false}
        },
        "required": ["leave_type_id", "start_date", "end_date"]
      },
...
    },
    "securitySchemes": {
      "sanctum": {
        "type": "apiKey",
        "in": "cookie",
        "name": "ogami_session",
        "description": "Laravel Sanctum session cookie authentication"
      }
    }
  }
}